backend/requirements.txt
fastapi==0.103.2
uvicorn==0.15.0
fastapi-cors
mysql-connector-python
black